pyproject.toml
[project]
name = "foundry-agents-mcp-server"
version = "0.1.0"
description = "MCP server for Azure AI Foundry agents, workflows, and vector search"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["mcp", "azure", "ai", "foundry", "agents", "search"]
dependencies = [
    "fastmcp>=2.0.0",
    "azure-ai-projects>=1.0.0b1",
    "azure-ai-agents>=1.0.0b1",
    "azure-search-documents>=11.5.0",
    "azure-identity>=1.15.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    # HTTP transport (Container Apps)
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    # Optional but imported with graceful fallback
    "azure-monitor-opentelemetry>=1.6.0",
    "opentelemetry-instrumentation-starlette>=0.50b0",
]

[project.optional-dependencies]
declarative = [
    "agent-framework-declarative>=0.1.0",
    "agent-framework-azure-ai>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/foundry_agents_mcp", "src/foundry_agents"]

[project.scripts]
foundry-agents-mcp-server = "foundry_agents_mcp.server:main"
deploy-case-study-agent = "foundry_agents.case_study_agent:deploy_cmd"
deploy-architecture-agent = "foundry_agents.architecture_agent:deploy_cmd"
run-project-log-workflow = "foundry_agents.project_log_workflow:run_cmd"
